requirements/client.txt
PyAudioWPatch
ffmpeg-python
scipy
websocket-client